+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New Features:
 
 * BGS-Tally now tracks carrier commodity buy and sell orders (in addition to the existing bartender materials buy and sell orders). These are tracked in real time as you change your carrier orders and are posted to Discord with your materials orders.
+* Adding a friend is now logged as a CMDR interaction.
 
 ### Changes: